Reorganizated connecter fieldScene mainAlg files
