+++ LARCmaCS/mainAlg.h
@@ -5,61 +5,16 @@
 #include <QDebug>
 #include <QSharedPointer>
 #include <QFile>
-#include "packetSSL.h"
 #include <iostream>
+
+#include "packetSSL.h"
 #include "mlData.h"
 #include "client.h"
+#include "mainAlgWorker.h"
 #define MAX_NUM_ROBOTS 12
 
 using namespace std;
 #include <time.h>       /* clock_t, clock(), CLOCKS_PER_SEC */
-
-struct MainAlgWorker : public QObject
-{
-    Q_OBJECT
-    clock_t timer,timer_s,timer_m,timer_max;
-    int Time_count;
-    bool Send2BT[MAX_NUM_ROBOTS];
-    Client client;
-    bool isPause;
-
-public:
-    MainAlgWorker();
-    ~MainAlgWorker() {}
-signals:
-    void sendToConnector(int N,QByteArray command);
-    void sendToBTtransmitter(char * message);
-    void mainAlgFree();
-    void StatusMessage(QString message);
-    void UpdatePauseState(QString message);
-
-public slots:
-
-    void start()
-    {
-        shutdowncomp = false;
-        cout << "MainAlg worker start" << endl;
-        init();
-    }
-
-    void Send2BTChangeit(bool *send2BT_);
-
-    void stop() { shutdowncomp = true; }
-    void run(PacketSSL packetssl);
-    void Pause();
-    void run_matlab();
-    void stop_matlab();    
-    void EvalString(QString s);
-
-private:
-    void init();
-
-    char m_buffer[256]; // matlab buffer
-    MlData fmldata;
-    bool fmtlab;
-    bool shutdowncomp;
-    bool pause;
-};
 
 struct MainAlg : public QObject
 {
@@ -69,32 +24,14 @@ public:
     MainAlgWorker worker;
     QThread thread;
 
-    explicit MainAlg(){}
-    ~MainAlg()
-    {
-        stop();
-        thread.wait(100);
-        thread.terminate();
-        thread.wait(100);
-    }
+    explicit MainAlg();
+    ~MainAlg();
 
-    void init()
-    {
-        worker.moveToThread(&thread);
-        cout << "Init mainAlg ok" << endl;
-        connect(this, SIGNAL(wstart()), &worker, SLOT(start()));
-        connect(this, SIGNAL(wstop()), &worker, SLOT(stop()));
-        connect(&thread, SIGNAL(finished()), &worker, SLOT(deleteLater()));
-    }
+    void init();
 
-    void start()
-    {
-        thread.start();
-        cout << "Thread start" << endl;
-        emit wstart();
-    }
+    void start();
 
-    void stop() { emit wstop(); }
+    void stop();
 
 signals:
     void wstart();
